Allow setting a query timeout on a per-query basis
Change `queryTimeout` to `defaultQueryTimeout` at the connection level
to make it clear it can be overridden.

The signature for `query` can be either a string or an object. The
`type` checking is the same as mysql2. It won't catch things like
passing an array through, but we wanted to match the underlying library.