Fix potential user issues in tests
Running as the host user ID fixed the npm install/tests when there were
already node modules, but errored with a cache error when there were no
node modules and the user didn't actually exist (e.g. GH Actions runs as
1001, which isn't a user inside the node container). This seems to
happen when the user's home directory does not exist.

The npm install needs to be done by the user that owns the directory. We
could add a Dockerfile and copy the files, but that seems overkill for
this!

Adding a new user with the host's user ID works in some circumstances,
but not if the host's user ID already exists inside the container (ie
1000). Changing the existing `node` user works in both cases.