Ensure lint container file are owned by host user
Same as the test container.

It's working fine locally but not in GHA:

npm ERR! Error: EACCES: permission denied, rename '/app/node_modules/@eslint-community/regexpp' -> '/app/node_modules/@eslint-community/.regexpp-MJHcVchs'

I think/guess it's because the lint and test containers share
node_modules (by volume) and npm can see that there is already stuff in
there that was installed by a different version of node, so is trying to
move it. All a bit gross, but hopefully this will sort it for now.